cmr-query/requirements.txt
python-cmr